requirements.txt
streamlit==1.42.2
numpy==1.26.4
pillow==9.4.0
tensorflow==2.18.0